Create an intermediate object for tracking publish requests
This adds a PublishRequest model for tracking the accessions and
where/how they should be published. This makes it easier to separate the
publishing process into multiple steps: publish and optionally
unembargo. This will be necessary going forward as the unembargo process
is going to get more complicated with the introduction of sponsored
bucket storage and IPTC tagging.